Patch book

+++ main.py
@@ -17,7 +17,12 @@ with app.app_context():
     db.create_all()
 
 bootstrap = Bootstrap5(app)
-
+def get_book(id):
+    try:
+        return db.get_or_404(Book, id)
+    except NotFound:
+        flash(f"Book with id {id} not found!")
+        return None
 @app.route("/books")
 def index():
     books = db.session.query(Book).all()
@@ -27,7 +32,7 @@ def index():
 def create_book():
     form = AddBookForm()
     if form.validate_on_submit():
-        title = form.book_name.data
+        title = form.title.data
         author = form.author.data
         rating = int(form.rating.data)
         book =  Book(title=title, author=author, rating=rating)
@@ -35,30 +40,46 @@ def create_book():
         db.session.commit()
         flash(f"{title} added successfully!", "success")
         return redirect(url_for("book_detail", id=book.id))
-    return render_template('create_book.html', form=form)
-
+    return render_template('create_book.html', heading="Create Book",url_action=('create_book', 0),form=form)
+@app.route("/books/<int:id>/edit", methods=["GET","POST", ])
+def edit_book(id):
+    book = get_book(id)
+    if book:
+        form = AddBookForm(obj=book)
+        if form.validate_on_submit():
+            book.title = form.title.data
+            book.author = form.author.data
+            book.rating = int(form.rating.data)
+            db.session.commit()
+            flash(f"Book details edited successfully!")
+            return redirect(url_for('index'))
+        return render_template("create_book.html",heading="Edit Book",url_action=('edit_book', id), form=form)
+    else:
+        return render_template("error.html")
 @app.route("/books/<int:id>", methods=["GET"])
 def book_detail(id):
-    book = db.get_or_404(Book,id)
-    message = session.pop("message", None)
-    if message:
-        flash(message)
-    return render_template("book_detail.html",book=book)
+    book = get_book(id)
+    if book:
+        message = session.pop("message", None)
+        if message:
+            flash(message)
+        return render_template("book_detail.html",book=book)
+    return render_template('error.html')
 @app.route("/books/<int:id>/delete", methods=["GET","POST"])
 def delete_book(id):
-    try:
         book = db.get_or_404(Book, id)
-        try:
-            db.session.delete(book)
-            flash(f"{book.title} deleted!")
-            db.session.commit()
-            return redirect(url_for('index'))
-        except Exception as e:
-            flash(f"{e}")
-            db.session.rollback()
-            return redirect(url_for('index'))
-    except NotFound:
-        flash(f"Book with id {id} not found!")
-        return redirect(url_for("index"))
+        if book:
+            try:
+                db.session.delete(book)
+                flash(f"{book.title} deleted!")
+                db.session.commit()
+                return redirect(url_for('index'))
+            except Exception as e:
+                flash(f"{e}")
+                db.session.rollback()
+                return redirect(url_for('index'))
+        else:
+            return redirect(url_for("index"))
+
 if __name__ == "__main__":
     app.run(debug=True)
